Unification of complex buffers API

+++ worker/include/MemoryBuffer.hpp
@@ -10,7 +10,7 @@ class MemoryBuffer
 {
 public:
     virtual ~MemoryBuffer() = default;
-    virtual uint64_t GetSize() const = 0;
+    virtual size_t GetSize() const = 0;
     virtual uint8_t* GetData() = 0;
     virtual const uint8_t* GetData() const = 0;
     bool IsEmpty() const { return 0U == GetSize() || nullptr == GetData() ; }
